Some internal values in pci.c are now 8-bit, which should reduce the number of screw-ups; ET4000/W32p PCI writes now sanitize the address first; Removed excess logging related to the ET4000/W32p.

+++ src/pci.c
@@ -10,7 +10,7 @@
 void    (*pci_card_write[32])(int func, int addr, uint8_t val, void *priv);
 uint8_t  (*pci_card_read[32])(int func, int addr, void *priv);
 void           *pci_priv[32];
-static int pci_index, pci_func, pci_card, pci_bus, pci_enable, pci_key;
+static uint8_t pci_index pci_func, pci_card, pci_bus, pci_enable, pci_key;
 static int pci_min_card, pci_max_card;
 int pci_burst_time, pci_nonburst_time;
